+++ app/app/api/private-cloud/requests/[id]/route.ts
@@ -2,8 +2,10 @@ import { DecisionStatus, EventType } from '@prisma/client';
 import { z } from 'zod';
 import { GlobalRole } from '@/constants';
 import createApiHandler from '@/core/api-handler';
+import prisma from '@/core/prisma';
 import { OkResponse, UnauthorizedResponse } from '@/core/responses';
-import { createEvent, models } from '@/services/db';
+import { sendRequestCancellationEmails } from '@/services/ches/private-cloud';
+import { createEvent, models, privateCloudRequestDetailInclude } from '@/services/db';
 
 const pathParamSchema = z.object({
   id: z.string(),
@@ -35,22 +37,27 @@ export const PUT = apiHandler(async ({ pathParams, session }) => {
     return UnauthorizedResponse();
   }
 
-  await models.privateCloudRequest.update(
-    {
-      where: {
-        id,
-        decisionStatus: DecisionStatus.PENDING,
-        active: true,
-      },
-      data: {
-        decisionStatus: DecisionStatus.CANCELLED,
-        active: false,
-      },
+  const updated = await prisma.privateCloudRequest.update({
+    where: {
+      id,
+      decisionStatus: DecisionStatus.PENDING,
+      active: true,
     },
-    session,
-  );
+    data: {
+      active: false,
+      decisionStatus: DecisionStatus.CANCELLED,
+      cancelledAt: new Date(),
+      cancelledById: session.user.id,
+    },
+    include: privateCloudRequestDetailInclude,
+  });
 
-  await createEvent(EventType.CANCEL_PRIVATE_CLOUD_REQUEST, session.user.id, { requestId: id });
+  const decoratedRequest = await models.privateCloudRequest.decorate(updated, session, true);
+
+  await Promise.all([
+    sendRequestCancellationEmails(decoratedRequest, session.user.name),
+    createEvent(EventType.CANCEL_PRIVATE_CLOUD_REQUEST, session.user.id, { requestId: id }),
+  ]);
 
   return OkResponse(true);
 });