+++ app/app/private-cloud/products/(product)/[licencePlate]/edit/page.tsx
@@ -145,12 +145,6 @@ export default privateCloudProductEdit(({ session }) => {
     },
   ];
 
-  const isMyProduct = [
-    snap.currentProduct.projectOwnerId,
-    snap.currentProduct.primaryTechnicalLeadId,
-    snap.currentProduct.secondaryTechnicalLeadId,
-  ].includes(session?.user.id ?? '');
-
   return (
     <div>
       <FormProvider {...methods}>
@@ -177,9 +171,7 @@ export default privateCloudProductEdit(({ session }) => {
         </form>
       </FormProvider>
 
-      {isMyProduct && (
-        <SiloAccordion className="my-4" disabled={isDisabled} licencePlate={snap.currentProduct?.licencePlate} />
-      )}
+      <SiloAccordion className="my-4" product={snap.currentProduct} />
     </div>
   );
 });