+++ app/emails/private-cloud/TeamRequestCancellation.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { getRandomUser } from '@/helpers/mock-resources/core';
+import { createSamplePrivateCloudRequest } from '@/helpers/mock-resources/private-cloud-request';
+import { PrivateCloudRequestDetailDecorated } from '@/types/private-cloud';
+import Template from '../_templates/private-cloud/TeamRequestCancellation';
+
+export default function TeamRequestCancellation() {
+  const sampleRequest = createSamplePrivateCloudRequest();
+  const sampleUser = getRandomUser();
+  return <Template request={sampleRequest as PrivateCloudRequestDetailDecorated} requester={sampleUser.displayName} />;
+}