+++ app/emails/public-cloud/TeamRequestCancellation.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { getRandomUser } from '@/helpers/mock-resources/core';
+import { createSamplePublicCloudRequest } from '@/helpers/mock-resources/public-cloud-request';
+import { PublicCloudRequestDetailDecorated } from '@/types/public-cloud';
+import Template from '../_templates/public-cloud/TeamRequestCancellation';
+
+export default function TeamRequestCancellation() {
+  const sampleRequest = createSamplePublicCloudRequest();
+  const sampleUser = getRandomUser();
+  return <Template request={sampleRequest as PublicCloudRequestDetailDecorated} requester={sampleUser.displayName} />;
+}